Checkout to the correct branch in with_checkout.
Currently in with_checkout, we use git reset --hard to move to the
correct hash, but we don't move onto the correct branch at the same
time.  For builds that depend on the current branch (e.g. they record
the branch into the build artifact) this behavior results in an
incorrect build.

Fix this by adding Cmd.git_checkout_force, which runs git checkout -f,
and calling that from with_checkout before the git reset --hard.

Signed-off-by: Ewan Mellor <ewan@tarides.com>